build 28.1
The volumes on the AWS I3 instance are returning such large inode
numbers that they overflow the signed integer that I am using to store
file state in the cache. This causes the cache manager code to
immediately discard a new cache entry as invalid and return empty
results. Changed database columes for stat to BIGINT to handle.